reunite a wayward search test with it's family
